fix: conditionally swap the bytes when building debug_id
When we are reading a minidump of different endianness we need to swap
the bytes of the integer fields of the GUID to build the correct
debug_id.  This because the debug_id is the hex representation of the
big-endian GUID.

When reading a little-endian minidump this does not swap, but since
the bytes are then implicitly swapped by reading them as little-endian
we end up converting their big-endian representation to hex.